fix failing test, next attempt
the previous version of the tests could sporadically fail if inserting
data into the in-memory index caches failed. this is possible, because
inserting index entries into the in-memory cache is not guaranteed to
work. inserting into the cache will fail if:
- the global memory budget of the cache subsystem is exceeded
- the cache's underlying hash table is in a "migration" phase
it is thus rarely possible that the insertion of any index entries into
an in-memory cache fails for a certain amount of time.
the fix in this PR is to try again in case of failure. up to 3 attempts
are made so that the tests do no execute endlessly.