fix unstable test (#17931)
* fix unstable test

=== server_parameters ===
* Test "server_parameters"
    [FAILED]  tests/js/client/server_parameters/auto-refill-index-caches-on-modify-noncluster-on.js

      "testUpdateEdgeDisbled" failed: Error: at assertion #5003: {
  "writesExecuted" : 0,
  "writesIgnored" : 0,
  "scannedFull" : 0,
  "scannedIndex" : 5000,
  "cursorsCreated" : 1,
  "cursorsRearmed" : 4999,
  "cacheHits" : 1147,
  "cacheMisses" : 3853,
  "filtered" : 0,
  "httpRequests" : 0,
  "executionTime" : 0.02600592700036941,
  "peakMemoryUsage" : 98304
}: (1147) is not equal to (0)
(Error
    at assertEqual (/work/ArangoDB/js/common/modules/jsunity/jsunity.js:108:19)
    at runCheck (tests/js/client/server_parameters/auto-refill-index-caches-on-modify-noncluster-on.js:64:7)
    at Object.testUpdateEdgeDisbled (tests/js/client/server_parameters/auto-refill-index-caches-on-modify-noncluster-on.js:151:7)
    at Object.run (/work/ArangoDB/js/common/modules/jsunity/jsunity.js:552:27)

* port changes from devel

* fix failing test, next attempt

the previous version of the tests could sporadically fail if inserting
data into the in-memory index caches failed. this is possible, because
inserting index entries into the in-memory cache is not guaranteed to
work. inserting into the cache will fail if:
- the global memory budget of the cache subsystem is exceeded
- the cache's underlying hash table is in a "migration" phase
it is thus rarely possible that the insertion of any index entries into
an in-memory cache fails for a certain amount of time.
the fix in this PR is to try again in case of failure. up to 3 attempts
are made so that the tests do no execute endlessly.