include database properties in debug dumps
currently only the database name is included in debug dumps. however the
database properties can also be quite important, for example, if the
database was a OneShard database.
with this change, debug dumps now include the relevant database
properties. this is a change in the debug dump format, but it should not
matter. currently the database name/properties are only printed, but not
used for anything else.