Added simple anti-overwhelm protection for telemetrics API
count the number of requests to the telemetrics API performed by the
arangosh, and rate-limit them.
by default, we only let 3 requests from the arangosh through to the
telemetrics API every 2 hours.
any additional requests will be responded to by the API with HTTP 420.
we do this to not over-report telemetry data, and also to protect
servers from being overwhelmed by too many telemetrics API requests
from batch programs running in arangosh.
note that the counters for rate-limiting are only stored in RAM on
single servers and coordinators. they are not persisted, so after a
server restart the counters are back at 0.
additionally, we do not keep track of the value across different
coordinators. in a load-balanced environment, the arangosh can
actually get up to (number of coordinators * maxRequestsPerBucket)
requests through in every interval.
we don't think this is a real problem that would justify a much more
complicated and less efficient request tracking (and probably some
coordination between servers).