fix single-to-single replication with authentication
single-to-single replication that used HTTP authentication to
authenticate requests on the leader could be broken if the collections
on the leader were created with 3.8 or later, and thus used the Merkle
tree protocol to exchange different document revisions.
the revisions API was extended with a document prefetching capability
about a year ago, and this prefetching capability uses the
NetworkFeature to send the requests out to the leader. The
NetworkFeature does not support HTTP authentication however, so on a
leader that used HTTP authentication, the requests were rejected with
HTTP 401 or HTTP 403, and replication failed.
when using JWTs, everything worked fine, but the single-to-single
replication also support HTTP authentication with different credentials
per database.