Fix race in kill queries test. (#18645)
The first server to simulate the query abortion reports back to the
coordinator, which will then send abort requests to all db servers.
These abort requests then race with the previous use requests. If the abort
request is processed before the use request, the assertions we had in
the code did not hold.