fix Snappy build on hosts that have BMI2 instructions
we unconditionally turn off BMI2 in this fix, because we cannot assume
BMI2 instructions to be present on all systems (minimum architecture
requirement is sandy-bridge architecture).
so this change will effectively not change our release builds, but may
turn off BMI2 unconditionally on some developer machines.
this should not matter much as we are moving away from Snappy and
towards LZ4 anyway.