disable `parallelize-gather` optimizer rules for some cases
in which it can lead to "query id ... not found" errors.
this change reverts the behavior of the `parallelize-gather` optimizer
rule to the same state it had in 3.10. the rule will be disabled for
some queries for which it was previously enabled only in devel/3.11.
the rule will be enabled again for these queries with a separate PR that
will contain a fix for the "query id ... not found" error and reenable
the rule.