increase too short timeouts
when fetching collection counts or selectivity estimates, the request
timeout for system collections was just 5s. this can be too low if DB
servers are busy.
also increase timeout to the regular timeout value in case the functions
are ever called for a SmartGraph edge collection part (_local_...,
_from_..., _to_...) which may have been recognized as system collections
due to the leading `_` in their name.