track index selectivity estimates memory usage
this PR tracks the approximate memory usage by index selectivity
estimates that use the RocksDBCuckooIndexEstimator.
memory is tracked for queued insert and remove operations. currently,
the memory is tracked by index, but the values can later be added to a
global metric.
the PR also release the memory used by the estimator early when the
collection is dropped. previously, the memory was not released until the
server shutdown. this was bad when a lot of collections were created and
dropped.
still missing in this PR is making the total memory usage of estimators
available as a metric.