reduce memory usage for incremental sync replication (#18835)
* reduce memory usage for incremental sync replication

the incremental sync protocol that was used for collections created with
3.8 or higher had memory usage issues in case the follower already had
some local data and its dataset was much larger than the leader's. For
example, this can happen if a follower gets disconnected, then a lot of
document removals happen on the leader and afterwards the follower tries
to get back into sync.

in this case, the follower buffered the ids of documents it had to
remove locally in a vector, which could grow arbitrarily large. the
removal of the documents contained in this vector would only happen at
the end, potentially even without performing intermediate commits.

the change in this PR is to trigger the document removal earlier, once
the vector has reached some size threshold, and also to use intermediate
commits during the removals.

* fix clobbering errors

* don't clear vectors twice

* apply review comments

---------

Co-authored-by: Vadim <vadim@arangodb.com>
Co-authored-by: CryptoNinjaGeek <86222482+CryptoNinjaGeek@users.noreply.github.com>