Fix TimeStampTransformer
Previously the implementation of TimeStampTransformer which takes a
std::chrono::system_clock::time_point and transforms it into a string
used the ArangoDB library function timepointToString.

This function in turn converted the time_point into a time_t and uses
gmtime to format the time_point.

If one uses std::chrono::system_clock::time_point::min() as input value
to timepointToString, this yields an invalid time_t on some
platforms (in particular here Windows), leading to an error in gmtime.

The ArangoDB library function ignores that error and continues.

This leads to invalid characters being left in the serialisation of the timepoint.

This PR changes the TimeStampTransformer to use formatting and parsing
functions from the date.h library to perform (de)serialisation of
timestamps.

It also adds some tests.