avoid recursive locking in agency
when loading configuration with an AQL query (which acquires the
collection's status lock) and then starting a transaction on the same
collection (acquires the status lock once more) while the query object
is still in scope. this leads to the same thread locking the same rw
lock in read mode twice, which can be dangerous if a writer is queued
for the lock. then the two threads will deadlock.