backport a change from devel
that relaxes a comparison made in the cluster maintenance for computed
values. not having computed values is stored as `null` normally, but
under some condition not having computed values can also be stored as an
empty array in devel. both values should be treated equal by the
maintenance. if the maintenance treats `null` and `[]` as different, it
may initiate a busy loops, trying forever to adjust the `computedValues`
property of the shard.
although we believe that in 3.11 having no computed values is always
stoted as `null`, this backport is made to avoid potential issues during
upgrades or in case other changes are backported from devel.
the added test should run into an issue in case the problem from devel
occurs.