fix wrong underlying parameter type (#19530)
for `--rocksdb.recycle-log-file-num`. The option does not make any
difference to use with the default WAL recovery mode we are using.
RocksDB automatically sanitizes the option to a value of `0`.

Previously we exposed this startup option as a boolean value, but it
should rather be a size_t. Interestingly, the RocksDB code also uses the
option with a wrong type in at least one place of their code.