added experimental backtrace API
by sending the arangod process a SIGUSR1 signal, it will make all its
threads log a backtrace. it does so by enumerating the list of current
threads via reading the fds in `/proc/<pid>/task` and using the `tgkill`
syscall to deliver a SIGUSR2 signal to each thread.
when a thread receives a SIGUSR2 signal, it will log its own backtrace
to the log.

this functionality is currently only available on Linux. it is highly
experimental and may not work in some environments (e.g. libmusl
builds).