harden callback functions against exceptions
these callback functions are called from C code which cannot handle
exceptions in any way. instead, we need to return HPE_INTERNAL (or
whatever other error) to the C code to signal that an error occurred.

also replaced previous occurrences of `return 0` which `return HPE_OK`.
`HPE_OK` is defined to `0` in llhttp.h, and is more telling than `return
0`.