remove SerializationFormat for AqlItemBlocks
we previously had 2 serialization formats, CLASSIC and SHADOWROWS. the
CLASSIC format was used up to ArangoDB 3.5 and was replaced with the
SHADOWROWS format then. for the transitioning time we offered support
for both formats. however, we already removed all support code for the
CLASSIC format a few releases ago, so it was effectively unusable for a
longer time already.
this PR now also removes all remainders in the code that passed the
format in the internal APIs. we are now always using the SHADOWROWS
format.