Backport #19768
Moves the arangotools target into the CMakeLists.txt in the directory
client-tools and adds the install statements for icudt.dat and the
common and client javascript code to the CMakeLists.txt inside
client-tools.

The behaviour I hope to achieve with this is that one can make -C
client-tools install DESTDIR=$somedir and gets a functioning set of
client tools.