make {Plain,Transactional}Cache::freeMemoryWhile() cheaper
the previous implementation of the function called `getBucket()` for
each slot it looked at. `getBucket()` then acquired a shared pointer on
the underlying table, which is not lock-free on x86_64.
instead, the calls to `Cache::table()` are guarded via a
pthread_mutex_lock and _unlock, which has prohibitive costs when called
too often.
the code is now changed so that there is a `getBucket()` overload which
gets the Table pointer from the outside, avoiding repeated calls to
acquire the Table pointer from `freeMemoryWhile()`.