add cluster-internal connectivity checks
This makes Coordinators and DB-Servers in a cluster periodically send check
requests to each other, in order to check if all servers can connect to
each other.
If a cluster-internal connection to another Coordinator or DB-Server cannot
be established within 10 seconds, a warning will be logged.

The new startup option `--cluster.connectivity-check-interval` can be used
to control the frequency of the connectivity check, in seconds.
If set to a value greater than zero, the initial connectivity check is
performed 15 seconds after the instance start, and subsequent connectivity
checks are executed with the specified frequency.
If set to a value of zero, connectivity checks are disabled.