fix potential recursive lock invocation (#19854)
this only happened if the cache manager was destroyed while caches were
still in use. in this case, the Manager's dtor called
Manager::shutdown(), which went through all still-available caches, and
called shutdown() on them. The shutdown would call
Manager::unregisterCache(), which could acquire the Manager _lock, and
then remove the Cache from the map of caches (_caches). If the last
shared_ptr to the Cache was destroyed, the Cache's dtor could try to
call into the Manager for a memory allocation ajdusted. this could
against acquire the Manager's _lock.