avoid duplicate creation of edge AqlValues (#19894)
inside single-server- or DB-server-based traversals.
the previous version of the code always built the AqlValue container for
the edge data twice, both when no projections were used and when
projections were used. when projections were used, the result value was
always reverted to the full edge, e.g. the projections optimization was
not used.