allow compiling executables with & without V8/JavaScript support
This introduces a CMake option `USE_V8`, which can be used to compile
our executables with and without V8/JavaScript support.
The option defaults to ON, so that V8/JavaScript are used as usual.
By setting the option to OFF, executables will be compiled without
V8/JavaScript support.

Note that V8 will always be built, regardless of the value of this option.
This is required because the V8 build step also builds ICU, which is needed
in all programs.
If set to OFF, V8 will not be linked to arangod and the client-tools,
and V8 integration code for arangod will not be compiled.
Additionally, arangosh will not be built.
Setting the option to OFF will further disable all JavaScript support in
arangod, rendering the web UI, JavaScript transactions, AQL user-defined
functions (UDFs), Foxx and some other smaller features unusable.
Setting this option to OFF also renders many tests unusable as many tests
depend on JavaScript.
Setting the option to OFF is highly experimental option and not well tested.
Use at your own risk!