make cache metrics more reliable
the previous implementation that returned the metrics for the in-memory
cache subsystem could return all metrics set to 0 if it could not
acquire the cache manager lock in time.
now, instead of returning zeroed results, we simply return the same
result that we returned last time. so now in the worst case we return
slightly stale metrics, but this is a lot better than returning zeroed
metrics.