silence "expected" errors
when a follower tries to get in sync with the leader and creates a lock
on the leader, the lock acquisition on the leader can fail. in this case
it is expected that a request to cancel the lock will always fail, with
the leader return HTTP_NOT_FOUND. this is an expected error in case the
lock wasn't acquired successfully, so there is no need to log it on the
follower as a warning.