unify handling of QueryOptions
in QueryOptions.cpp, limit the scope of the lookup result for each query
option tested, and use `isTrue()` instead of `getBool()` for boolean
values. the latter provides a tiny speedup compared to `getBool()`.
additionally, unify the behavior between the handling of the C++
QueryOptions.allPlans attribute, which only picks up boolean values, and
its JavaScript counterpart, which only checked for truth. this could
lead to JavaScript errors in case a trueish, but non-boolean value was
given for `allPlans` via the JavaScript code, e.g.
```
db._query(query, bindVars, {allPlans: "foo"});
```