Set build-docker-image default value in base config
If a pipeline parameter is set explicitly, this value will implicitly be
forwarded to the continuation pipeline. If a value is not explicitly set.
but has a default value, then this value is _not_ implicitly forwarded.
It can be explicitly forwarded by using the continuation's parameters,
but if we do that unconditionally, this causes a problem once the parameter
is again set explicitly, because then we get a "Conflicting pipeline parameters."
error because the parameter is forwarded twice (once implicitly and once explicitly).