Fix handling of non existing properties in Maintenance
Previously, if the planned slice did not contain a property, we simply
ignored it, even if we still had that property in local. Now we always
consider this property as null for the comparison.
However, for replication2 we must not compare WriteConcern and
WaitForSync, because there these properties are managed differently.