execute shell_client test in forked arangosh
if the shell_client test is executed in the same arangosh instance as
the basic test executor, then all suppressions that are set by the basic
test executor for the sanitizers (TSan, ASan) will not be used by the
arangosh instance itself. the suppressions will only be picked up
correctly by newly started processes. thus we need to execute the
shell_client test in a new arangosh instance.