[BTS-1636] Fix path found by SHORTEST_PATH with weight is not path of lowest weight (#20305)
When searching for (in particular weighted) shortest paths, ArangoDB searches forward from the start and backward from the target vertex with something that looks quite a lot like a Dijkstra search.

An important aspect is when a search is finished, that is when a lowest weight path has been found. Whenever a vertex is relaxed, every newly discovered vertex is checked whether it has been seen on the other side of the search. If it has been a path has been found, but there is no guarantee that this is the lowest weight one.

The search has to continue until the sum of the diameters, the weight of the highest proven weight paths, exceeds the weight of the best candidate found.

This patch implements this terminating condition.

In addition, since the two sided enumerator supports enumerating all paths between two vertices in increasing order by weight, by extension all weights below the sum of the diameters threshold are deemed to be optimal and hence all paths with weight below the threshold can be returned.