fix non-deterministic behavior on Windows
the register planning could insert different variables with the same
register id into the VarInfoMap.
the info from the VarInfoMap was later copied into a vector. VarInfoMap
is an unordered container with no guaranteed iteration order. thus it
was not deterministic which of the inserted variables with the same
register id value was found first.
the issue did not occur in any of the Linux builds, potentially because
there the iteration order for unordered containers is more deterministic
than on Windows, for better or worse. it was a bug anyway, and the
iteration order must not play a role here.