Implement SupervisedScheduler::detachThread [BTS-1670]
This is one measure to resolve https://arangodb.atlassian.net/browse/BTS-1670

In this PR we create a possiblity that a SchedulerThread can notice that
it might be up to some longer running task (like waiting for a lock) and
then voluntarily decide to detach itself from the scheduler to make room
for a new thread to avoid starvation of jobs on the queue.

For now, we use this for acquiring a collection level lock (as read lock
for write operations and as write lock for exclusive access), if we
cannot acquire the lock within 1s. This will then allow the scheduler
to create new threads to continue work on whatever priority is suitable.