lower memory usage high water mark multiplier for in-memory cache
to 56% of the configured cache size value.
these 56% are the same value that was used in previous version of
ArangoDB. we temporarily increased the value in devel, but it turned out
that the cache can still overcommit memory because its memory
reclamation procedure is asynchronous and can run in parallel to other
tasks that insert new data.
to stay on the safe side and not overcommit more memory usage than
before, the default high water multiplier value is reverted to 56%, the
same value as used in all previous versions.