Fix issue that prefetch tasks could flood the scheduler queue (#20573)
* Fix issue that prefetch tasks could flood the scheduler queue

Async prefetching could post a task onto the scheduler. If this task is not picked up in time, we could claim that task for ourselves and simply to the work directly. However, we could post yet another task for the next batch. This could go on and on, filling up the scheduler queue. As a side effect, there was also a race where an old task could be executed with an old AqlCallStack.
To avoid this, we now rearm the already posted task with the current AqlCallStack and set it to pending. This way we can ensure that either the updated task is available to be picked up, or if it was picked up before we could make it available again, we will realize and can then repost it on the scheduler.

* Add static_assert to ensure _state is always lock-free