improve performance of constrained sort executor
as the final step in the constrained sort executor, the min/max heap is
sorted to get a totally ordered result back.
the previous implementation used std::sort() for this, which is over the
top. as we have already built up a heap, we can use std::sort_heap() to
do the sorting. this can be algorithmically more efficient.