improve debuggability of collection-counts test (#20595)
also increase the timeouts for waitForShardsInSync in this test from 60
seconds to the default value (300s). 60 seconds may not be enough if the
cluster is under really high load (it was in this situation when the
latest failure occurred in this test case).
furthermore, log a more telling error message in case the shard
distribution cannot be retrieved properly. previously we only logged an
info that the shard distribution could not be retrieved. now we are also
logging the actual response message we got. this may help us figure out
what actually went wrong.
finally, add the selected follower servers to the supervision message
for a successful addFollower job execution.