fix sporadic query setup errors (#20653)
during testing there were sporadic errors when setting up AQL queries
via the route POST `/_api/aql/setup`:

    AQL: continued non-continuable method for POST /_api/aql/setup

this error is due to setupClusterQuery() returning a future, which is
being waited for using waitForFuture(). waitForFuture() returns
RestStatus::DONE if the future was already executed, and everything is
fine then.
if the future was not already executed, waitForFuture() returns
RestStatus::WAITING, so that eventually the RestHandler is restarted via
its continueExecute() method. For the RestAqlHandler, this always
triggered the above error if restarting was done for the HTTP POST setup
request.
the fix is to allow restarting the RestHandler in case of HTTP POST as
well, and then simply return RestStatus::DONE.