do not move uncorrelated subqueries down
via optimizer rule `move-calculations-down`.
doing so can cause performance degradations during query plan
optimization for no good reason.

for example, if there is the following query

  FOR doc IN ...
    LET result1 = (...)
    LET result2 = (...)
    ...

and `result1` and `result2` are uncorrelated, then there is no benefit
in moving `result1` down past `result2`. doing so anyway will trigger a
recalculation of which variables are valid in the plan at any step,
which is expensive and thus we want to avoid doing this unnecessarily.