remove `_activeTrx` instance variable
the `RestDocumentHandler` had an instance variable `_activeTrx` in which
the handler stored the transaction object that was started for the
particular operation.
this was previously necessary, because the transaction still needs to be
valid in some of the RestDocumentHandler's follow-up async
continuations. because we did not pass the unique_ptr with the
transaction to these async routines, they needed the instance variable
to pick up the transaction later on.

after we changed the `RestDocumentHandler` to use coroutines, we can now
remove the `_activeTrx` instance variable entirely, and simply store the
transaction in a local variable. once the coroutine execution resumes,
it still has access to that same local variable.