obsolete 3.11 feature flag (#20814)
this PR adds the option `--query.enable-debug-apis` as an obsolete
option in devel.
the option was introduced in 3.11.9 as actual feature flag for some
debugging APIs. these APIs are always available in devel, and cannot be
turned off.
so while setting this option has no effect in devel, the existence of
the options allows users to upgrade from ArangoDB 3.11.9 with having the
startup option explicitly set.
if devel would not recognize the startup option as a valid option, it
would otherwise bail out with an error on startup, thus preventing a
successful upgrade.