make optimizer rule `async-prefetch` unavailable in 3.11. (#20823)
make the optimizer rule `async-prefetch` fully unavailable in 3.11, so
that enabling it by accident does not cause any queries to execute the
rule.
the motivation for this change is that although the `async-prefetch`
optimizer rule already existed in 3.11, it was disabled by default and
can lead to errors in query execution when enabled. thus it is better to
fully disable the optimizer rule in 3.11 and only bring it back with
version 3.12.0, where the rule's code has been rewritten and fixed.