BTS-1489: Race condition in AsioSocket shutdown when using SSL. (#20815)
All AsioSocket instances are now created via `std::make_shared()`. This
allows us to pass a shared_ptr to an AsioSocket into a lambda, so the
lambda can still safely refer to the AsioSocket without any lifetime
issues.
With the fix it is possible that the AsioSocket posts a lambda with a
this pointer (to the AsioSocket), and the lambda being executed after
the AsioSocket is destroyed.