fix protocol confusion (#20900)
when an async job was created via the jobs API via HTTP/1.1, the job
result was stored as an HttpResponse. When the job result was retrieved
via HTTP/2, the server code expected the response to be an H2Response (a
struct that is derived from HttpResponse). this could lead to the wrong
response class to be used, and thus potential UB. the situation in which
this happens is rare though, because it requires usage of the jobs API
and a protocol change.