prevent parallel index creation thread from hanging
indefinitely, in case the thread objects were created successfully but
starting a thread returned an error.
in this case, the threads that were unable to start would never call
the completion function on the shared env, so that the main index
creation thread would wait forever for all threads to complete.
now, if a thread cannot be started successfully, we immediately count
down the number of threads to wait for in the shared env.