make latest indexes visible (#20941)
when calling GET /_api/index?withHidden=true.
we need to wait here for the latest index information being present in
the agency cache. if we don't wait, we risk not seeing a just-created
index if the local agency cache has not fully caught up, and the
supervision declares the just-created index ready.