turn an ADB_PROD_ASSERT into an exception
it is much better to make the server process throw a meaningful
exception rather than make it crash when invalid user input is sent to
the database.

in this case, it was possible to send the following data to the create
database API and trigger a production assertion:
```
"{\"name\":\"OneShard\",\"users\":[],\"options\":{\"replicationFactor\":\"satellite\",\"sharding\":\"single\",\"writeConcern\":3}}"
```
the problem here is that the replicationFactor "satellite" and using a
writeConcern seem to be mutually exclusive. the same is true for using
replicationFactor "satellite" in a OneShard database.