try to fix shutdown errors
it seems that under some unknown circumstances the Boost asio
io_contexts can still execute actions after the destructors of the
NetworkFeature ran. this PR is an attempt to improve the situation, by
stopping and clearing all available io_context objects when shutting
down one of the ConnectionPool objects.
it is unclear yet if this will fix the issue or not.