BTS-1813: disable failing subquery chaos test
this PR disables the subquery chaos test that oftentimes fails in our
nightly CI runs. it does not fix the assertion failure the test runs
into.
this assertion failure still needs to be addressed properly. there is
ticket BTS-1813 for this, which will remain open until the issue is
fixed.