BTS-1775: disable checking of view counts in chaos test
view counts are known to differ from collection counts when the
DeleteRange version of truncate is used. this is noted in BTS-1775.
until there is a proper fix for the view counts when using DeleteRange,
we should disable the count check in the chaos test, in order to avoid
spurious failed tests for known reasons