try to fix more shutdown errors (#20959)
* try to fix more shutdown errors

AsyncAgencyComm requests may still be in flight when the server is
shutting down. Try to fix this.

* try to fix shutdown errors

it seems that under some unknown circumstances the Boost asio
io_contexts can still execute actions after the destructors of the
NetworkFeature ran. this PR is an attempt to improve the situation, by
stopping and clearing all available io_context objects when shutting
down one of the ConnectionPool objects.
it is unclear yet if this will fix the issue or not.

* simplify PR

---------

Co-authored-by: Vadim Kondratev <vadim@arangodb.com>