avoid dropping of followers in some case
avoid dropping of followers in the following scenario:
- leader has in-sync follower(s) for a given shard
- resignLeadership is called for current leader
- resignLeadership is reverted, so that old leader is now again the leader
- write to the shard

previously this led to dropping the followers when the write operation
was executed, because the leader's session id changed when taking over
shard leadership again. the new session id was propagated to all
followers, but it wasn't picked up because the leader server id did not
change.