honor special case "OFF" for ARANGODB_BUILD_DATE cmake var
by default, CMake variables default to "OFF". this case was not handled
separately for the `ARANGODB_BUILD_DATE` cmake variable. we only
compared the value of this variable to the empty string.
this PR adjusts the cmake code to also take into account the special
value of "OFF", so that the build date for an ArangoDB executable will
not be set to "OFF" if the build date is not explicitly set via the
`ARANGODB_BUILD_DATE` cmake variable.