refresh collection count value asynchronously (#21003)
* refresh collection count value asynchronously

via the scheduler. this improves latency of count() operations that use
the cache if the cache value was previously set but is now expired. in
this case, the thread that asked for the count() will be returned the
expired stale count value from the cache, and a refresh of the cache
value is ordered to be executed eventually via the scheduler, using the
LOW priority queue.

* make scheduled cache refresh operation async as well

* Make the count-update code a little cleaner with a coroutine

* Added a comment

---------

Co-authored-by: Tobias Gödderz <tobias@arangodb.com>