avoid dropping of followers in some case (#21006)
* avoid dropping of followers in some case

avoid dropping of followers in the following scenario:
- leader has in-sync follower(s) for a given shard
- resignLeadership is called for current leader
- resignLeadership is reverted, so that old leader is now again the leader
- write to the shard

previously this led to dropping the followers when the write operation
was executed, because the leader's session id changed when taking over
shard leadership again. the new session id was propagated to all
followers, but it wasn't picked up because the leader server id did not
change.

* added test

* adjust for devel

* avoid potential races during controlled leader change

this PR adds a new method `setTheLeaderConditional` to FollowerInfo, so
that the leader only gets changed if the current leader is still as
expected. this avoids potential races if concurrent leader changes are
going on.