fix a potential Merkle tree corruption issue
potential data corruption in a collection's Merkle tree in case a
write operation in a streaming transaction ran into the transaction's
maximum size limit and failed. if the transaction was continued
afterwards and more operations were pushed into it, the leftovers of the
failed operation were not removed from the collection's in-memory
buffers that contained all modifications for the collection's Merkle
tree upon commit of that transaction.

sequence of events needed:
1. streaming transaction
2. write operation in streaming transaction ran into transaction size
   limit error
3. the write operation error was ignored by the user
4. the transaction was committed anyway