fix a potential use-after-free error
the `calculationVocbase` object is a `TRI_vocbase_t` instance that can
outlive the lifetime of the application server in case
`DatabaseFeature::unprepare()` is not called. in this case, the
calculationVocbase's dtor is called only by the executable's static
finalizers. this must be prevented, as `TRI_vocbase_t` contains other
objects, which upon destruction may call into other ApplicationFeatures,
e.g. the MetricsFeature.