fix potential use-after-free in replication
the variable `requestPayload` was only present inside a certain block of
code, and would be destroyed when leaving the code block. however, tasks
could have been scheduled that still referred to this variable by
reference, so the tasks referring to the variable could outlive the
lifetime of the variable itself.
this is now fixed by moving the `requestPayload` variable outside of the
scope, so it is guaranteed to stay valid until the very end of the
scope.