fix data race in FollowerInfo
`FollowerInfo::newShardEntry()` could read instance variables without
acquiring the RW lock when called from `FollowerInfo::add()`. this could
lead to a situation in which a thread in `newShardEntry` reads the
instance variables without synchronization, while another thread
modifies the instance variables (and correctly using the RW lock in
write mode when doing so).