make sure environment variable names are unique
when starting an external process, we pass a list of environment
variables to the system call that starts the process.
the names of environment variables are not guaranteed to be unique, so
when an environment variable is specified multiple times, it is not
guaranteed that the last occurrence wins. this is problematic in case we
pass additional environment variables to the process, which should
always win over whatever settings were originally in the environment.